Get boss to shoot the projectiles
